Update everything to use new main extensions
