+++ clikt/src/commonMain/kotlin/com/github/ajalt/clikt/completion/CompletionGenerator.kt
@@ -1,10 +1,11 @@
 package com.github.ajalt.clikt.completion
 
+import com.github.ajalt.clikt.core.BaseCliktCommand
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.PrintCompletionMessage
 
 internal object CompletionGenerator {
-    fun throwCompletionMessage(command: CliktCommand, shell: String): Nothing {
+    fun throwCompletionMessage(command: BaseCliktCommand<*>, shell: String): Nothing {
         val message = when (shell.trim().lowercase()) {
             "fish" -> FishCompletionGenerator.generateFishCompletion(command = command)
             "zsh" -> BashCompletionGenerator.generateBashOrZshCompletion(
