+++ samples/copy/src/main/kotlin/com/github/ajalt/clikt/samples/copy/main.kt
@@ -1,6 +1,7 @@
 package com.github.ajalt.clikt.samples.copy
 
 import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.core.main
 import com.github.ajalt.clikt.core.terminal
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.multiple
