+++ samples/json/src/main/kotlin/com/github/ajalt/clikt/samples/json/main.kt
@@ -2,6 +2,7 @@ package com.github.ajalt.clikt.samples.json
 
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.context
+import com.github.ajalt.clikt.core.main
 import com.github.ajalt.clikt.core.subcommands
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.multiple
