+++ samples/repo/src/main/kotlin/com/github/ajalt/clikt/samples/repo/main.kt
@@ -1,6 +1,7 @@
 package com.github.ajalt.clikt.samples.repo
 
 import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.core.main
 import com.github.ajalt.clikt.core.requireObject
 import com.github.ajalt.clikt.core.subcommands
 import com.github.ajalt.clikt.output.TermUi
