+++ samples/validation/src/main/kotlin/com/github/ajalt/clikt/samples/validation/main.kt
@@ -1,6 +1,7 @@
 package com.github.ajalt.clikt.samples.validation
 
 import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.core.main
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.convert
 import com.github.ajalt.clikt.parameters.options.*
